[refactor] 옵저버 패턴에서 옵저빙되는 객체, 옵저버의 종류에 맞게 패키지 분리? ?
