+++ src/design/observer/practice/subject/Message.java
@@ -1,4 +1,4 @@
-package design.observer.practice;
+package design.observer.practice.subject;
 
 import java.time.LocalDateTime;
