+++ src/design/observer/practice/subject/Reader.java
@@ -1,4 +1,4 @@
-package design.observer.practice;
+package design.observer.practice.subject;
 
 public interface Reader {
     void update();
